// web/shared/api/types.gen.ts
export type BodyUploadProfileImageProfileMeImageTypePost = {
    /**
     * Image
     */
    image: string;
};
// ...
export type UserSchema = {
    /**
     * Id
     */
    id: number;
    /**
     * Username
     */
    username: string;
    /**
     * Email
     */
    email: string;
    /**
     * Avatar
     */
    avatar?: string;
    /**
     * Banner
     */
    banner?: string;
};